started to include bootstrap
replaced vertical own created navbar with a horizontal one
predesigned from bootstrap.

for this to work, I had to use a <div> and fill the
content with a JavaScript on page load.
Further more, for Sites NOT on the SiteRoot, I included
a base tag into the head, because the links would be
broken instead.
Possible that absolute paths would work from the server,
but locally called as file in the browser, a absolute path
isn't working as intended.

Next step to-do now is to read about the grid system from
bootstrap to get the content layouted fine for all kinds
of displays